Adding arguments to /start
I tried to make /start command compatible with arguments. In this commit, the `CommandHandler('start')` has been allowed to accept arguments, and the `download_command` has been expanded to check how many arguments there are.
- If the length of the `args` array is 0, meaning there are no arguments, the /start function will behave as usual.
- If the length of the `args` array is 1, meaning there is 1 argument, the /start function will trigger the sending of the proper langpack via the `platform_handler` function.
- If the length of the `args` array is more than 1, the bot will send an error message.

The `platform_handler` function has been also expanded in order to filter whether it was called by a Callback_Query or by the `/start ARGUMENT` function. This modification is not still perfect and it will not work until the `/start` string is deleted from `message.text`, because the bot will receive `/start ARGUMENT`. So you need to find how to get rid of the first `/start ` text.